pyproject.toml
[project]
name = "bank-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crispy-bootstrap5>=2025.6",
    "django>=5.2.6",
    "django-crispy-forms>=2.4",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "pillow>=11.3.0",
    "psycopg>=3.2.10",
    "shortuuid>=1.0.13",
]
